[bugfix] pos-amounts
