+++ apps/backend/src/flat-files/tdi34/TDI34Details.ts
@@ -19,7 +19,7 @@ export interface ITDI34Details extends IFixedWidthRecord<ITDI34Details> {
   transaction_date: string;
   transaction_time: string;
   settlement_date: string;
-  transaction_code: TransactionCode;
+  transaction_code: number;
   fill2: string;
   approval_code: string;
   fill3: string;
@@ -132,7 +132,7 @@ export class TDI34Details
     this.resource.settlement_date = data;
   }
 
-  @Column({ start: 67, width: 2 })
+  @Column({ start: 67, width: 2, format: { type: DataType.Integer } })
   public get transaction_code() {
     return this.resource.transaction_code;
   }
@@ -178,12 +178,13 @@ export class TDI34Details
 
   public set transaction_amt(data) {
     if (
-      this.transaction_code === TransactionCode.MERCH_RETURN ||
-      this.transaction_code === TransactionCode.PURCHASE_ADJUSTMENT
+      this.resource.transaction_code === TransactionCode.MERCH_RETURN ||
+      this.resource.transaction_code === TransactionCode.PURCHASE_ADJUSTMENT
     ) {
-      data = data * -1;
+      this.resource.transaction_amt = data * -1;
+    } else {
+      this.resource.transaction_amt = data;
     }
-    this.resource.transaction_amt = data;
   }
 
   @Column({ start: 89, width: 10 })
